+++ lang/derive/accounts/src/lib.rs
@@ -432,7 +432,7 @@ use syn::parse_macro_input;
 ///                 <br><br>
 ///                 Example:
 ///                 <pre>
-/// use anchor_spl::{mint, token::{TokenAccount, Mint}};
+/// use anchor_spl::{mint, token::{TokenAccount, Mint, Token}};
 /// ...&#10;
 /// #[account(
 ///     init,
@@ -445,6 +445,8 @@ use syn::parse_macro_input;
 /// pub mint: Account<'info, Mint>,
 /// #[account(mut)]
 /// pub payer: Signer<'info>,
+/// pub token_program: Program<'info, Token>,
+/// pub system_program: Program<'info, System>
 ///                 </pre>
 ///             </td>
 ///         </tr>
@@ -461,7 +463,7 @@ use syn::parse_macro_input;
 ///                 <br><br>
 ///                 Example:
 ///                 <pre>
-/// use anchor_spl::token::Mint;
+/// use anchor_spl::token::{Mint, Token};
 /// ...&#10;
 /// #[account(
 ///     init,
@@ -480,6 +482,8 @@ use syn::parse_macro_input;
 /// pub mint_two: Account<'info, Mint>,
 /// #[account(mut)]
 /// pub payer: Signer<'info>,
+/// pub token_program: Program<'info, Token>,
+/// pub system_program: Program<'info, System>
 ///                 </pre>
 ///             </td>
 ///         </tr>
@@ -493,7 +497,11 @@ use syn::parse_macro_input;
 ///                 <br><br>
 ///                 Example:
 ///                 <pre>
-/// use anchor_spl::{mint, token::{TokenAccount, Mint}};
+/// use anchor_spl::{
+///     associated_token::AssociatedToken,
+///     mint,
+///     token::{TokenAccount, Mint, Token}
+/// };
 /// ...&#10;
 /// #[account(
 ///     init,
@@ -511,6 +519,9 @@ use syn::parse_macro_input;
 /// pub mint: Account<'info, Mint>,
 /// #[account(mut)]
 /// pub payer: Signer<'info>,
+/// pub token_program: Program<'info, Token>,
+/// pub associated_token_program: Program<'info, AssociatedToken>,
+/// pub system_program: Program<'info, System>
 ///                 </pre>
 ///             </td>
 ///         </tr>