+++ lang/src/accounts/program.rs
@@ -162,12 +162,6 @@ impl<'info, T: Id + Clone> ToAccountInfos<'info> for Program<'info, T> {
     }
 }
 
-impl<'info, T: Id + Clone> ToAccountInfo<'info> for Program<'info, T> {
-    fn to_account_info(&self) -> AccountInfo<'info> {
-        self.info.clone()
-    }
-}
-
 impl<'info, T: Id + Clone> AsRef<AccountInfo<'info>> for Program<'info, T> {
     fn as_ref(&self) -> &AccountInfo<'info> {
         &self.info
@@ -182,9 +176,4 @@ impl<'info, T: Id + Clone> Deref for Program<'info, T> {
     }
 }
 
-impl<'info, T: Id + Clone> AccountsExit<'info> for Program<'info, T> {
-    fn exit(&self, _program_id: &Pubkey) -> ProgramResult {
-        // No-op.
-        Ok(())
-    }
-}
+impl<'info, T: AccountDeserialize + Id + Clone> AccountsExit<'info> for Program<'info, T> {}