+++ ts/src/program/namespace/account.ts
@@ -1,7 +1,5 @@
-import { Buffer } from "buffer";
 import camelCase from "camelcase";
 import EventEmitter from "eventemitter3";
-import bs58 from "bs58";
 import {
   Signer,
   PublicKey,
@@ -13,11 +11,7 @@ import {
 } from "@solana/web3.js";
 import Provider, { getProvider } from "../../provider.js";
 import { Idl, IdlTypeDef } from "../../idl.js";
-import Coder, {
-  ACCOUNT_DISCRIMINATOR_SIZE,
-  accountSize,
-  AccountsCoder,
-} from "../../coder/index.js";
+import { Coder, BorshCoder } from "../../coder/index.js";
 import { Subscription, Address, translateAddress } from "../common.js";
 import { AllAccountsMap, IdlTypes, TypeDef } from "./types.js";
 import * as pubkeyUtil from "../../utils/pubkey.js";
@@ -126,9 +120,8 @@ export class AccountClient<
     this._idlAccount = idlAccount;
     this._programId = programId;
     this._provider = provider ?? getProvider();
-    this._coder = coder ?? new Coder(idl);
-    this._size =
-      ACCOUNT_DISCRIMINATOR_SIZE + (accountSize(idl, idlAccount) ?? 0);
+    this._coder = coder ?? new BorshCoder(idl);
+    this._size = this._coder.accounts.size(idlAccount);
   }
 
   /**
@@ -144,15 +137,6 @@ export class AccountClient<
     if (accountInfo === null) {
       return null;
     }
-
-    // Assert the account discriminator is correct.
-    const discriminator = AccountsCoder.accountDiscriminator(
-      this._idlAccount.name
-    );
-    if (discriminator.compare(accountInfo.data.slice(0, 8))) {
-      throw new Error("Invalid account discriminator");
-    }
-
     return this._coder.accounts.decode<T>(
       this._idlAccount.name,
       accountInfo.data
@@ -188,15 +172,9 @@ export class AccountClient<
       commitment
     );
 
-    const discriminator = AccountsCoder.accountDiscriminator(
-      this._idlAccount.name
-    );
     // Decode accounts where discriminator is correct, null otherwise
     return accounts.map((account) => {
       if (account == null) {
-        return null;
-      }
-      if (discriminator.compare(account?.account.data.slice(0, 8))) {
         return null;
       }
       return this._coder.accounts.decode(
@@ -223,24 +201,16 @@ export class AccountClient<
   async all(
     filters?: Buffer | GetProgramAccountsFilter[]
   ): Promise<ProgramAccount<T>[]> {
-    const discriminator = AccountsCoder.accountDiscriminator(
-      this._idlAccount.name
-    );
-
     let resp = await this._provider.connection.getProgramAccounts(
       this._programId,
       {
         commitment: this._provider.connection.commitment,
         filters: [
           {
-            memcmp: {
-              offset: 0,
-              bytes: bs58.encode(
-                filters instanceof Buffer
-                  ? Buffer.concat([discriminator, filters])
-                  : discriminator
-              ),
-            },
+            memcmp: this.coder.accounts.memcmp(
+              this._idlAccount.name,
+              filters instanceof Buffer ? filters : undefined
+            ),
           },
           ...(Array.isArray(filters) ? filters : []),
         ],