+++ CHANGELOG.md
@@ -11,12 +11,20 @@ incremented for features.
 
 ## [Unreleased]
 
-### Fixes
+## [0.20.1] - 2022-01-09
 
-### Features
+### Fixes
 
 * lang: Improved error msgs when required programs are missing when using the `init` constraint([#1257](https://github.com/project-serum/anchor/pull/1257))
 * lang: Add `seeds::program` constraint for specifying which program_id to use when deriving PDAs.([#1197](https://github.com/project-serum/anchor/pull/1197))
+
+### Features
+
+* lang: Allow repr overrides for zero copy accounts ([#1273](https://github.com/project-serum/anchor/pull/1273)).
+
+### Breaking
+
+* ts: `Coder` is now an interface and the existing class has been renamed to `BorshCoder`. This change allows the generation of Anchor clients for non anchor programs  ([#1259](https://github.com/project-serum/anchor/pull/1259/files)).
 
 ## [0.20.0] - 2022-01-06